requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
black==22.10.0
click==8.1.3
colorama==0.4.6
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.1.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-registration==3.3
django-registration-redux==2.11
executing==1.1.1
flake8==5.0.4
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==3.20.1
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
traitlets==5.5.0
tzdata==2022.5
wcwidth==0.2.5
